requirements.txt
asgiref==3.5.2
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
Django==4.1
django-debug-toolbar==3.6.0
django-dotenv==1.4.2
django-filter==22.1
django-request-logging==0.7.5
djangorestframework==3.13.1
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
wincertstore==0.2
